# API/FeatherBasic/public/dist/swagger4.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API接口"
  },
  "host": "127.0.0.1:3030",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/VideoCount/RequeArray": {
      "post": {
        "tags": [
          "VideoCount"
        ],
        "summary": "传入一个用户号码数组",
        "description": "",
        "operationId": "createUsersWithArrayInput",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "用户号码数组",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Phone"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful opertation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/VIDEORESP"
              }
            }
          },
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/PhoneOn": {
      "post": {
        "tags": [
          "ControllPhone"
        ],
        "summary": "传入一个用户号码数组",
        "description": "",
        "operationId": "ControllPhoneWithArrayInput",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "用户号码数组",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Phone"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful opertation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/VIDEORESP"
              }
            }
          },
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/PhoneOn/{phone}": {
      "get": {
        "tags": [
          "ControllPhone"
        ],
        "summary": "终端设置开关机状态查询",
        "description": "",
        "operationId": "Phoneget",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "x-api-key",
            "in": "query",
            "description": "需要填入授权信息",
            "required": true,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "path",
            "description": "需要填入号码",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "put": {
        "tags": [
          "ControllPhone"
        ],
        "summary": "终端开机状态更改",
        "description": "",
        "operationId": "PhoneOn",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "x-api-key",
            "in": "query",
            "description": "需要填入授权信息",
            "required": true,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "path",
            "description": "需要填入号码",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "ControllPhone"
        ],
        "summary": "终端关机",
        "description": "",
        "operationId": "PhoneOff",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-api-key",
            "in": "query",
            "description": "需要填入授权信息",
            "required": true,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "path",
            "description": "需要填入号码",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/JumpMessage": {
      "get": {
        "tags": [
          "QueryJump"
        ],
        "summary": "终端注册信息查询",
        "description": "",
        "operationId": "QueryJumpInfoty",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "x-api-key",
            "in": "query",
            "description": "需要填入授权信息",
            "required": true,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "path",
            "description": "需要填入号码",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/EmitKKST": {
      "post": {
        "tags": [
          "EmitKKST"
        ],
        "summary": "康卡斯特激活",
        "description": "",
        "operationId": "EmitKKSTsss",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "用户号码数组",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Phone"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful opertation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/VIDEORESP"
              }
            }
          },
          "default": {
            "description": "successful operation"
          }
        }
      }
    }
  },
  "definitions": {
    "Phone": {
      "type": "object",
      "properties": {
        "phone": {
          "$ref": "#/definitions/PhoneCatGory"
        },
        "x-api-key": {
          "type": "string"
        }
      }
    },
    "PhoneCatGory": {
      "type": "array",
      "xml": {
        "name": "phonenumber",
        "wrapped": true
      },
      "items": {
        "type": "string"
      }
    },
...
    "VIDEORESP": {
      "type": "object",
      "properties": {
        "phone": {
          "type": "string"
        },
        "devid": {
          "type": "string"
        }
      }
    },
...
  },
...
}